fix(bench shard): address code review findings
- Make --stratify-by repo actually affect the partition: with the flag,
  instances are grouped by repo and shuffled per-group so every repo is
  spread evenly across shards; without it, a single uniform global shuffle
  is used. Previously the flag changed only manifest metadata.
- Avoid cloning the whole instance vector in run_shard; capture the source
  ID set and total count, then move instances into the partitioner.
- Compute per-shard counts and balance spread once; reuse for both the
  pre-write invariant check and the returned report.
- Validate --shards >= 1 at the clap layer (value_parser range) so 0 fails
  fast with a standard message.
- Always record stratify_mode in the manifest, since the mode affects
  leftover placement even when --stratify-by is off.
- Update spec-shard.md to match the new --stratify-by semantics and add
  regression tests for even repo spread and mode recording.

+++ src/cli/args.rs
@@ -1720,8 +1720,8 @@ pub struct ShardCmd {
     pub dataset_cache_dir: Option<PathBuf>,
 
     /// Number of output shards.  Must be ≥ 1 and ≤ the dataset instance count.
-    #[arg(long, value_name = "N")]
-    pub shards: usize,
+    #[arg(long, value_name = "N", value_parser = clap::value_parser!(u64).range(1..))]
+    pub shards: u64,
 
     /// Destination directory for shard files.  Created if absent; must be empty otherwise.
     #[arg(long)]
